optimize duplication for poollist
