+++ src/pages/Beta/Pool/AllPoolList/PoolList/PoolListV2.tsx
@@ -1,6 +1,4 @@
-import React, { useCallback, useEffect, useState, useContext, memo } from 'react'
-import { ThemeContext } from 'styled-components'
-import { TextInput, Box, Loader } from '@pangolindex/components'
+import React, { useCallback, useEffect, useState, memo } from 'react'
 import {
   MinichefStakingInfo,
   useSortFarmAprs,
@@ -12,32 +10,11 @@ import {
 import { DOUBLE_SIDE_STAKING_REWARDS_INFO } from 'src/state/stake/doubleSideConfig'
 import PoolCardV2 from '../PoolCard/PoolCardV2'
 import { useChainId } from 'src/hooks'
-import { useTranslation } from 'react-i18next'
-import { Search } from 'react-feather'
 import useDebounce from 'src/hooks/useDebounce'
 import { BIG_INT_ZERO } from 'src/constants'
-import Scrollbars from 'react-custom-scrollbars'
-import { PoolsWrapper, PanelWrapper, LoadingWrapper, MobileGridContainer } from './styleds'
 import { usePoolDetailnModalToggle } from 'src/state/application/hooks'
-import DetailModal from '../../DetailModal'
-import DropdownMenu from 'src/components/Beta/DropdownMenu'
-import { Hidden } from 'src/theme'
-
-export enum SortingType {
-  totalStakedInUsd = 'totalStakedInUsd',
-  totalApr = 'totalApr'
-}
-
-export const SortOptions = [
-  {
-    label: 'Liquidity',
-    value: SortingType.totalStakedInUsd
-  },
-  {
-    label: 'APR',
-    value: SortingType.totalApr
-  }
-]
+import { SortingType } from './PoolCardListView'
+import PoolCardListView from './PoolCardListView'
 
 export interface EarnProps {
   version: string
@@ -50,8 +27,7 @@ export interface EarnProps {
 
 const PoolListV2: React.FC<EarnProps> = ({ version, stakingInfos, setMenu, activeMenu, menuItems }) => {
   const chainId = useChainId()
-  const { t } = useTranslation()
-  const theme = useContext(ThemeContext)
+
   const [poolCardsLoading] = useState(false)
 
   const [searchQuery, setSearchQuery] = useState<string>('')
@@ -120,85 +96,34 @@ const PoolListV2: React.FC<EarnProps> = ({ version, stakingInfos, setMenu, activ
   )
   const selectedPool = selectedPoolIndex !== -1 ? stakingInfoData[selectedPoolIndex] : ({} as MinichefStakingInfo)
 
-  const renderPoolCardListView = () => {
-    if ((stakingRewardsExist && stakingInfoData?.length === 0) || poolCardsLoading)
-      return (
-        <LoadingWrapper>
-          <Loader size={100} />
-        </LoadingWrapper>
-      )
-    else if ((!stakingRewardsExist || stakingInfoData?.length === 0) && !poolCardsLoading) {
-      return <div>{t('earnPage.noActiveRewards')}</div>
-    } else {
-      return (
-        <>
-          <Box display="flex" justifyContent="space-between" alignItems="center" mb={10}>
-            <Box width="100%">
-              <TextInput
-                placeholder={t('searchModal.tokenName')}
-                onChange={handleSearch}
-                value={searchQuery}
-                id="token-search-input"
-                addonAfter={<Search style={{ marginTop: '5px' }} color={theme.text2} size={20} />}
-              />
-            </Box>
-            <Hidden upToSmall={true}>
-              <Box ml={10}>
-                <DropdownMenu
-                  title="Sort by:"
-                  options={SortOptions}
-                  value={sortBy}
-                  onSelect={value => {
-                    setSortBy(value)
-                  }}
-                  height="54px"
-                />
-              </Box>
-            </Hidden>
-          </Box>
-          <MobileGridContainer>
-            <DropdownMenu
-              options={menuItems}
-              value={activeMenu}
-              onSelect={value => {
-                setMenu(value)
-              }}
-            />
-            <DropdownMenu
-              title="Sort by:"
-              options={SortOptions}
-              value={sortBy}
-              onSelect={value => {
-                setSortBy(value)
-              }}
-            />
-          </MobileGridContainer>
-
-          <Scrollbars>
-            <PanelWrapper>
-              {stakingInfoData.map((stakingInfo, index) => (
-                <PoolCardV2
-                  key={stakingInfo?.pid}
-                  stakingInfo={stakingInfo}
-                  onClickViewDetail={() => {
-                    setSelectedPoolIndex(index)
-                    togglePoolDetailModal()
-                  }}
-                  version={Number(version)}
-                />
-              ))}
-            </PanelWrapper>
-          </Scrollbars>
-        </>
-      )
-    }
-  }
-
   return (
-    <PoolsWrapper>
-      {renderPoolCardListView()}
-      <DetailModal stakingInfo={selectedPool} version={Number(version)} />
-    </PoolsWrapper>
+    <PoolCardListView
+      version={version}
+      setMenu={value => setMenu(value)}
+      activeMenu={activeMenu}
+      menuItems={menuItems}
+      handleSearch={handleSearch}
+      onChangeSortBy={value => {
+        setSortBy(value)
+      }}
+      sortBy={sortBy}
+      searchQuery={searchQuery}
+      isLoading={(stakingRewardsExist && stakingInfoData?.length === 0) || poolCardsLoading}
+      isNotExitsPool={(!stakingRewardsExist || stakingInfoData?.length === 0) && !poolCardsLoading}
+      selectedPool={selectedPool}
+    >
+      {stakingInfoData.map((stakingInfo, index) => (
+        <PoolCardV2
+          key={stakingInfo?.pid}
+          stakingInfo={stakingInfo}
+          onClickViewDetail={() => {
+            setSelectedPoolIndex(index)
+            togglePoolDetailModal()
+          }}
+          version={Number(version)}
+        />
+      ))}
+    </PoolCardListView>
   )
 }
